+++ src/reduxStore/slices/accountInfoSlice.ts
@@ -4,8 +4,6 @@ import { REHYDRATE } from 'redux-persist';
 import { ZERO } from 'constants/index';
 import { BatchTransactionsWSResponseType } from 'types';
 import { AccountType } from 'types/account.types';
-import { storage } from 'utils/storage';
-import { localStorageKeys } from 'utils/storage/local';
 
 import {
   loginAction,
@@ -162,7 +160,6 @@ export const accountInfoSlice = createSlice({
   },
   extraReducers: (builder) => {
     builder.addCase(logoutAction, () => {
-      storage.local.removeItem(localStorageKeys.loginExpiresAt);
       return initialState;
     });
     builder.addCase(