+++ src/reduxStore/slices/loginInfoSlice.ts
@@ -1,6 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { IframeLoginTypes } from 'lib/sdkWebWalletIframeProvider';
-import { setLoginExpiresAt, getNewLoginExpiresTimestamp } from 'storage/local';
 import { LoginMethodsEnum } from 'types/enums.types';
 import { TokenLoginType } from '../../types';
 import {
@@ -22,7 +21,6 @@ export interface LedgerLoginType {
 
 export interface LoginInfoType {
   data: any;
-  expires?: number;
 }
 
 export interface LoginInfoStateType {
@@ -135,7 +133,6 @@ export const loginInfoSlice = createSlice({
         state.isLoginSessionInvalid = false;
         state.loginMethod = action.payload.loginMethod;
         state.iframeLoginType = action.payload.iframeLoginType;
-        setLoginExpiresAt(getNewLoginExpiresTimestamp());
       }
     );
   }